Break out render-test utils
