+++ docs/api-reference/test-utils/test-renderer.md
@@ -1,0 +1,14 @@
+# TestRenderer (Test Automation Class, Experimental)
+
+A customized [`SceneRenderer`](./docs/api-reference/test-utils/node-test-driver.md).
+
+## Test Case Format
+
+* `name`
+* `views` (defaults to `[new MapView()]`)
+* `viewState`
+* `layers`
+* `referenceImage`
+
+
+## Methods
