+++ examples/test-utils/README.md
@@ -1,0 +1,3 @@
+# deck.gl/test-utils
+
+This folder contains files that the user may want to copy to set up their own tests.
