+++ examples/test-utils/index.html
@@ -1,0 +1,11 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset='UTF-8' />
+    <title>rendering test</title>
+	<link href="app.css" rel="stylesheet">
+  </head>
+  <body>
+    <script src='test-rendering-bundle.js'></script>
+  </body>
+</html>
