+++ package.json
@@ -35,15 +35,18 @@
     "publish-prod": "npm run build && npm run test && npm run test-dist && npm publish",
     "publish-beta": "npm run build && npm run test && npm run test-dist && npm publish --tag beta",
     "test": "npm run lint && npm run build && npm run test-node && (cd test/render && node node-rendering.js)",
+    "test2": "npm run lint && npm run build && npm run test-node && npm run test-render-puppet",
     "test-fast": "npm run test-node",
     "test-ci": "npm run lint && node test/node-ci.js",
     "test-cover": "NODE_ENV=test tape -r babel-register test/node.js && nyc report",
     "test-node": "node test/node.js",
     "test-dist": "node test/node-dist.js",
     "test-browser": "webpack-dev-server --env.test --progress --hot --open",
+    "test-render": "node test/render-from-node.js",
+    "test-render-puppet": "node test/render-from-node.js",
+    "test-render-browser": "webpack-dev-server --env.render --progress --hot --open",
     "bench": "node test/bench/node.js",
-    "bench-browser": "webpack-dev-server --env.bench --progress --hot --open",
-    "test-rendering": "(cd test/render && webpack-dev-server --config webpack.config.test-rendering.js --progress --hot --open)"
+    "bench-browser": "webpack-dev-server --env.bench --progress --hot --open"
   },
   "dependencies": {
     "d3-hexbin": "^0.2.1",
