+++ src/test-utils/src/index.js
@@ -1,6 +1,17 @@
-export * from './layer-lifecycle-utils';
-export {toLowPrecision} from './precision';
-export {default as spy} from './spy';
-export {default as makeSpy} from './spy';
-export {default as gl} from './setup-gl';
+export {toLowPrecision} from './utils/precision';
+export {default as spy} from './utils/spy';
+export {default as makeSpy} from './utils/spy';
+export {default as gl} from './utils/setup-gl';
 export * from '../luma.gl/gpgpu';
+
+// Basic utility for rendering multiple scenes (could go into "deck.gl/core")
+export {default as SceneRenderer} from './scene-renderer';
+
+// Utilities for update tests (lifecycle tests)
+export * from './lifecycle-test';
+// A utility that renders a list of scenes and compares against golden images
+export {default as RenderTest} from './render-test';
+
+// Node.js test drivers
+export {default as NodeTestDriver} from './drivers/node-test-driver';
+export {default as RenderTestDriver} from './drivers/render-test-driver';
